# openapi.yaml
openapi: 3.0.3
info:
  title: Chat Widget Auth & History API
  description: >
    채팅 위젯의 **인증(Auth)** 및 **대화 내역 저장(History Storage)**을 담당하는 백엔드 API입니다.
    
    **도메인 정책 (v1.3):**
    - 위젯 키 생성 시 `https://` 프로토콜을 제외한 도메인만 입력받습니다.
    - 루트 도메인, 서브 도메인 와일드카드 정책이 적용됩니다.
  version: 1.3.1

servers:
  - url: http://localhost:3000

tags:
  - name: Widget Auth
    description: (Public) 위젯 초기화 및 세션 발급
  - name: Widget Messages
    description: (Public) 대화 내역 저장 및 조회
  - name: Admin Management
    description: (Private) 위젯 키 관리

paths:
  # ----------------------------
  # 1. Widget Authentication
  # ----------------------------
  /api/v1/widget/auth/session:
    post:
      tags:
        - Widget Auth
      summary: 위젯 세션 토큰 발급
      description: >
        위젯 키와 현재 페이지의 도메인을 검증하여 **세션 토큰**을 발급합니다.
        
        **검증 로직 상세:**
        1. 요청 헤더의 `Origin`이 존재하면 이를 우선 검증값으로 사용합니다.
        2. `Origin`이 없는 경우(일부 환경) body의 `pageUrl`에서 도메인을 추출합니다.
        3. 두 값이 모두 존재하는데 서로 다르다면 요청은 거부(Spoofing 의심)됩니다.
        4. 추출된 도메인(프로토콜 제거)이 DB의 `allowedDomains` 규칙과 매칭되는지 확인합니다.
        5. 해당 `widgetKey`가 `REVOKED` 상태라면 발급을 거부합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WidgetSessionRequest'
      responses:
        '200':
          description: 세션 발급 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WidgetSessionResponse'
        '400':
          description: 잘못된 요청 (Origin/PageUrl 불일치 등)
        '403':
          description: 도메인 검증 실패 또는 REVOKED 된 키
        '404':
          description: 존재하지 않는 widgetKey

  # ----------------------------
  # 2. Chat History Storage
  # ----------------------------
  /api/v1/widget/messages:
    get:
      tags:
        - Widget Messages
      summary: 대화 내역 조회 (페이징)
      security:
        - widgetSessionAuth: []
      parameters:
        - in: query
          name: cursor
          schema:
            type: string
          description: "이전 페이지의 마지막 메시지 ID (없으면 최신순 조회)"
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          description: "한 번에 가져올 메시지 개수"
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatMessage'
                  nextCursor:  # <--- [중요] 추가됨: 다음 페이지 조회를 위한 커서
                    type: string
                    nullable: true
                    description: "다음 메시지를 불러오기 위한 커서 ID (null이면 더 이상 없음)"

    post:
      tags:
        - Widget Messages
      summary: 대화 메시지 저장
      security:
        - widgetSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageInput'
      responses:
        '201':
          description: 저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessage'

  # ----------------------------
  # 3. Admin Management
  # ----------------------------
  /api/v1/admin/widget-keys:
    get:
      tags:
        - Admin Management
      summary: 위젯 키 목록 조회
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WidgetKey'

    post:
      tags:
        - Admin Management
      summary: 위젯 키 생성
      description: >
        새로운 위젯 키를 생성합니다. 생성된 직후 `secretKey`가 반환됩니다.
        
        **도메인 등록 규칙:**
        - 프로토콜(`https://`)은 제외하고 입력하세요.
        - `*.example.com` 와일드카드 지원.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, allowedDomains]
              properties:
                name:
                  type: string
                  example: "메인 쇼핑몰용"
                allowedDomains:
                  type: array
                  items:
                    type: string
                    example: "*.myshop.com"
      responses:
        '201':
          description: 생성 성공
          content:  # <--- [중요] 추가됨: 생성된 키 정보를 반환해야 함
            application/json:
              schema:
                $ref: '#/components/schemas/WidgetKey'

  /api/v1/admin/widget-keys/{widgetKeyId}/revoke:
    patch:
      tags:
        - Admin Management
      summary: 위젯 키 폐기 (Revoke)
      description: >
        특정 키를 `REVOKED` 상태로 변경하여 더 이상 세션 발급이 불가능하게 만듭니다.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: widgetKeyId
          required: true
          schema:
            type: string
            format: uuid
          description: "폐기할 위젯 키의 UUID"
      responses:
        '200':
          description: 폐기 성공 (상태 변경됨)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WidgetKey'
        '404':
          description: 존재하지 않는 Key ID

# ----------------------------
# Components
# ----------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "관리자용 토큰 (Admin Dashboard)"
    widgetSessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "위젯 세션 토큰 (Widget Client)"

  schemas:
    # --- Auth ---
    WidgetSessionRequest:
      type: object
      required: [widgetKey, pageUrl]
      properties:
        widgetKey:
          type: string
          example: "wk_live_abc123"
        pageUrl:
          type: string
          description: "위젯이 실행된 현재 페이지 URL (Origin 검증 보조용)"
          example: "https://www.myshop.com/products/1"

    WidgetSessionResponse:
      type: object
      properties:
        sessionToken:
          type: string
          description: "JWT Access Token"
        expiresIn:
          type: integer
          description: "토큰 만료 시간(초)"

    # --- Chat ---
    ChatMessageInput:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        metadata:
          type: object
          description: "토큰 사용량 등 부가 정보"

    ChatMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    # --- Admin ---
    WidgetKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        secretKey:
          type: string
          example: "wk_live_xyz789"
        status:
          type: string
          enum: [ACTIVE, REVOKED]
        allowedDomains:
          type: array
          items:
            type: string
          example: ["myshop.com", "*.myshop.com"]
        createdAt:
          type: string
          format: date-time